+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/main.js
@@ -17,8 +17,20 @@ define([
     'sulucontact/model/termsOfDelivery',
     'sulucontact/model/email',
     'sulucontact/model/emailType',
-    'sulumedia/model/media'
-], function(Account, Contact, AccountContact, AccountsUtilHeader, Activity, TermsOfPayment, TermsOfDelivery, Email, EmailType, Media) {
+    'sulumedia/model/media',
+    'accountsutil/delete-dialog'
+], function(
+    Account,
+    Contact,
+    AccountContact,
+    AccountsUtilHeader,
+    Activity,
+    TermsOfPayment,
+    TermsOfDelivery,
+    Email,
+    EmailType,
+    Media,
+    DeleteDialog) {
 
     'use strict';
 
@@ -664,18 +676,7 @@ define([
 
         // show confirmation and delete account
         del: function() {
-            this.confirmSingleDeleteDialog(this.options.id, function(wasConfirmed, removeContacts) {
-                if (wasConfirmed) {
-                    this.sandbox.emit('sulu.header.toolbar.item.loading', 'options-button');
-                    this.account.destroy({
-                        data: {removeContacts: !!removeContacts},
-                        processData: true,
-                        success: function() {
-                            this.sandbox.emit('sulu.router.navigate', 'contacts/accounts');
-                        }.bind(this)
-                    });
-                }
-            }.bind(this));
+            DeleteDialog.showForSingle(this.sandbox, this.account, this.options.id);
         },
 
         // saves an account
@@ -753,22 +754,7 @@ define([
                 this.sandbox.emit('sulu.overlay.show-error', 'sulu.overlay.delete-no-items');
                 return;
             }
-            this.showDeleteConfirmation(ids, function(wasConfirmed, removeContacts) {
-                if (wasConfirmed) {
-                    // TODO: show loading icon
-                    ids.forEach(function(id) {
-                        var account = new Account({id: id});
-                        account.destroy({
-                            data: {removeContacts: !!removeContacts},
-                            processData: true,
-
-                            success: function() {
-                                this.sandbox.emit('husky.datagrid.record.remove', id);
-                            }.bind(this)
-                        });
-                    }.bind(this));
-                }
-            }.bind(this));
+            this.showDeleteConfirmation(ids);
         },
 
         renderList: function() {
@@ -860,173 +846,18 @@ define([
             return dfd.promise();
         },
 
-        showDeleteConfirmation: function(ids, callbackFunction) {
+        showDeleteConfirmation: function(ids) {
             if (ids.length === 0) {
                 return;
             } else if (ids.length === 1) {
                 // if only one account was selected - get related sub-companies and contacts (and show the first 3 ones)
-                this.confirmSingleDeleteDialog(ids[0], callbackFunction);
+                //this.confirmSingleDeleteDialog(ids[0], callbackFunction);
+                DeleteDialog.showForSingle(this.sandbox, Account.findOrCreate({id:ids[0]}), ids[0], true)
             } else {
                 // if multiple accounts were selected, get related sub-companies and show simplified message
-                this.confirmMultipleDeleteDialog(ids, callbackFunction);
-            }
-        },
-
-        confirmSingleDeleteDialog: function(id, callbackFunction) {
-            var url = '/admin/api/accounts/' + id + '/deleteinfo';
-
-            this.sandbox.util.ajax({
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-
-                context: this,
-                type: 'GET',
-                url: url,
-
-                success: function(response) {
-                    this.showConfirmSingleDeleteDialog(response, id, callbackFunction);
-                }.bind(this),
-
-                error: function(jqXHR, textStatus, errorThrown) {
-                    this.sandbox.logger.error("error during get request: " + textStatus, errorThrown);
-                }.bind(this)
-            });
-        },
-
-        showConfirmSingleDeleteDialog: function(values, id, callbackFunction) {
-            // check if callback is a function
-            if (!!callbackFunction && typeof(callbackFunction) !== 'function') {
-                throw 'callback is not a function';
-            }
-
-            var content = 'contact.accounts.delete.desc',
-                furtherChildren,
-                furtherContacts,
-                overlayType = 'show-warning',
-                title = 'sulu.overlay.be-careful',
-                okCallback = function() {
-                    var deleteContacts = this.sandbox.dom.find('#overlay-checkbox').length && this.sandbox.dom.prop('#overlay-checkbox', 'checked');
-                    callbackFunction.call(this, true, deleteContacts);
-                }.bind(this);
-
-            // sub-account exists => deletion is not allowed
-            if (parseInt(values.numChildren, 10) > 0) {
-
-                furtherChildren = values.numChildren - values.children.length;
-                overlayType = 'show-error';
-                title = 'sulu.overlay.error';
-                okCallback = undefined;
-                // parse sub-account template
-                content = this.sandbox.util.template(templates.dialogEntityFoundTemplate, {
-                    foundMessage: this.sandbox.translate('contact.accounts.delete.sub-found'),
-                    list: this.template.dependencyListAccounts.call(this, values.children),
-                    numChildren: parseInt(values.numChildren, 10),
-                    andMore: this.sandbox.util.template(this.sandbox.translate('public.and-number-more'), {number: furtherChildren}),
-                    description: this.sandbox.translate('contact.accounts.delete.sub-found-desc')
-                });
-            }
-            // related contacts exist => show checkbox
-            else if (parseInt(values.numContacts, 10) > 0) {
-
-                furtherContacts = values.numContacts - values.contacts.length;
-
-                // create message
-                content = this.sandbox.util.template(templates.dialogEntityFoundTemplate, {
-                    foundMessage: this.sandbox.translate('contact.accounts.delete.contacts-found'),
-                    list: this.template.dependencyListContacts.call(this, values.contacts),
-                    numChildren: parseInt(values.numContacts, 10),
-                    andMore: this.sandbox.util.template(this.sandbox.translate('public.and-number-more'), {number: furtherContacts}),
-                    description: this.sandbox.translate('contact.accounts.delete.contacts-question'),
-                    checkboxText: this.sandbox.util.template(this.sandbox.translate('contact.accounts.delete.contacts-checkbox'), {number: parseInt(values.numContacts, 10)})
-                });
-            }
-
-            // show dialog
-            this.sandbox.emit('sulu.overlay.' + overlayType,
-                title,
-                content,
-                callbackFunction.bind(this, false),
-                okCallback
-            );
-        },
-
-        confirmMultipleDeleteDialog: function(ids, callbackFunction) {
-            var url = '/admin/api/accounts/multipledeleteinfo';
-            this.sandbox.util.ajax({
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-
-                context: this,
-                type: 'GET',
-                url: url,
-                data: {ids: ids},
-
-                success: function(response) {
-                    this.showConfirmMultipleDeleteDialog(response, ids, callbackFunction);
-                }.bind(this),
-
-                error: function(jqXHR, textStatus, errorThrown) {
-                    this.sandbox.logger.error("error during get request: " + textStatus, errorThrown);
-                }.bind(this)
-            });
-        },
-
-        showConfirmMultipleDeleteDialog: function(values, ids, callbackFunction) {
-            // check if callback is a function
-            if (!!callbackFunction && typeof(callbackFunction) !== 'function') {
-                throw 'callback is not a function';
-            }
-
-            var content = 'contact.accounts.delete.desc',
-                title = 'sulu.overlay.be-careful',
-                overlayType = 'show-warning',
-                okCallback = function() {
-                    var deleteContacts = this.sandbox.dom.find('#delete-contacts').length && this.sandbox.dom.prop('#delete-contacts', 'checked');
-                    callbackFunction(true, deleteContacts);
-                }.bind(this);
-
-            // sub-account exists => deletion is not allowed
-            if (parseInt(values.numChildren, 10) > 0) {
-                overlayType = 'show-error';
-                title = 'sulu.overlay.error';
-                okCallback = undefined;
-                content = this.sandbox.util.template(templates.dialogEntityFoundTemplate, {
-                    foundMessage: this.sandbox.translate('contact.accounts.delete.sub-found'),
-                    description: this.sandbox.translate('contact.accounts.delete.sub-found-desc')
-                });
-            }
-            // related contacts exist => show checkbox
-            else if (parseInt(values.numContacts, 10) > 0) {
-                // create message
-                content = this.sandbox.util.template(templates.dialogEntityFoundTemplate, {
-                    foundMessage: this.sandbox.translate('contact.accounts.delete.contacts-found'),
-                    numChildren: parseInt(values.numContacts, 10),
-                    description: this.sandbox.translate('contact.accounts.delete.contacts-question'),
-                    checkboxText: this.sandbox.util.template(this.sandbox.translate('contact.accounts.delete.contacts-checkbox'), {number: parseInt(values.numContacts, 10)})
-                });
-            }
-
-            // show dialog
-            this.sandbox.emit('sulu.overlay.' + overlayType,
-                title,
-                content,
-                callbackFunction.bind(this, false),
-                okCallback
-            );
-        },
-
-        template: {
-            dependencyListContacts: function(contacts) {
-                var list = "<% _.each(contacts, function(contact) { %> <li><%= contact.firstName %> <%= contact.lastName %></li> <% }); %>";
-                return this.sandbox.template.parse(list, {contacts: contacts});
-            },
-            dependencyListAccounts: function(accounts) {
-                var list = "<% _.each(accounts, function(account) { %> <li><%= account.name %></li> <% }); %>";
-                return this.sandbox.template.parse(list, {accounts: accounts});
+                //this.confirmMultipleDeleteDialog(ids, callbackFunction);
+                DeleteDialog.showForMultiple(this.sandbox, ids);
             }
         }
-
     };
 });